Add Matching engine functionality

+++ src/main/java/service/OrderBookService.java
@@ -7,6 +7,6 @@ import java.math.BigDecimal;
 public interface OrderBookService {
 
     void addOrder(LimitOrder order);
-    void modifyOrder(Long orderId, int newQuantity, BigDecimal newPrice);
+    void modifyOrder(Long orderId, int newQuantity, int newPrice);
     void deleteOrder(Long orderId);
 }
